requirements-dev.txt
# requirements.txt
binance-connector==3.5.1
customtkinter==5.2.1
pandas==2.1.4
python-binance==1.0.19
python-dotenv==1.0.0
packaging
sqlite3
plotly


# requirements-dev.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==7.0.0
mypy==1.8.0
click==8.1.7
coverage==7.4.1